Mostly integrated w/ socketconnection and virtualserver
This commit is mostly debugged and more or less functional. However, when a player uses the rungame command in a lobby, memory corruption often occurs, and in-game commands from the players become garbled and unusable, occasionally resulting in segfaults.

From what I can tell, this is in some way related to the section in the Game.cpp constructor where usernames and passwords are pulled from text files. When in game, executing "cd ..", the raw player input (printed in line 45 of Player.cpp) reads as a random chunk of passwords from the passwords.txt file.